tests/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
docopt==0.6.2
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
paho-mqtt==1.6.1
pbr==6.1.0
pipdeptree==2.23.4
pluggy==1.5.0
PyJWT==2.9.0
pykwalify==1.8.0
pytest==8.3.3
pytest-django==4.9.0
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
sqlparse==0.5.1
stevedore==4.1.1
urllib3==2.2.3
